ocr-python-service/requirements.txt
Flask
pytesseract
Pillow
pdf2image
PyPDF2
pdfminer.six
camelot-py
